add asset pipeline

+++ Makefile
@@ -8,5 +8,11 @@ format:
 	${GOFMT} -w ${GOFILES}
 
 clean:
-	rm chatwit
+	if [ -f chatwit ] ; then rm chatwit ; fi
+	if [ -d public/assets ] ; then  rm -rf public/assets ; fi
+	if [ -f .*.css ] ; then  rm .*.css ; fi
+	if [ -f .*.js ] ; then  rm .*.js ; fi
+
+assets-bundle:
+	${GOPATH}/bin/train bundle
